Fix: 편지 뒤로 가기 버튼 삭제
